% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_attributenumber}
\alias{calculate_attributenumber}
\title{Generates the number of attributes of a data.frame object.}
\usage{
calculate_attributenumber(df)
}
\arguments{
\item{df}{A data.frame object.}
}
\value{
The number of attributes of \code{df}.
If the above input params are wrong, an error message is returned.
}
\description{
This function generates the number of attributes of a data.frame object.
As input is only required an object from the class data.frame.
The data.frame object should represent an multivariate ts.
Otherwise the function returns an error message.
}
\examples{
calculate_attributenumber(df = multi_ts_list$`M-TS-1`$data)
}
