% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_autocorrelation}
\alias{calculate_autocorrelation}
\title{Generates the mean of the autocorrelationfunction (ACF) of an ts object.}
\usage{
calculate_autocorrelation(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The mean of the ACF of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This is a function to generate the mean of the ACF of an time series object.
As input is only required an object from the class time series.
Otherwise the function returns an error message.
}
\examples{
calculate_autocorrelation(ts = tsobject)
}
