% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_determination_coefficient}
\alias{calculate_determination_coefficient}
\title{Generates the coefficient of determination (R2) of an data.frame object.}
\usage{
calculate_determination_coefficient(df, targetcol)
}
\arguments{
\item{df}{A data.frame object.}

\item{targetcol}{A string containing a column name of the \code{df}.}
}
\value{
The R2 of \code{df}.
If the input is not a data.frame, an error message is returned.
}
\description{
This is a function to generate the R2 of an data.frame object.
As input is only required an object from the class data.frame.
The data.frame object should represent an multivariate ts.
Otherwise the function returns an error message.
}
\examples{
calculate_determination_coefficient(df = dataframe_object,targetcol = 'var1')
}
