% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_dtw_blockdistance}
\alias{calculate_dtw_blockdistance}
\title{Generates the dynamic time warping (DTW) for an ts object to the 1000 ts
from the list in /data.}
\usage{
calculate_dtw_blockdistance(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The DTW of \code{ts} to the 10 different blocks. Thus an vector
with 10 DTW distances is returned. The first number is the DTW to
block 1,..., the last number is the DTW to block 10. If the input is not
a ts object, an error message is returned.
}
\description{
This is a function to generate the DTW for a ts to the 10 blocks from the
1000 ts in the list in /data. Each block contains 100 ts. As input is only
required an object from the class time series. Otherwise the function
returns an error message.
}
\examples{
calculate_dtw_blockdistance(ts = tsobject)
}
