% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_durbin_watson_test}
\alias{calculate_durbin_watson_test}
\title{Generates the autocorrelation measure of an data.frame object.}
\usage{
calculate_durbin_watson_test(df, targetcol)
}
\arguments{
\item{df}{A data.frame object.}

\item{targetcol}{A string containing a column name of the \code{df}.}
}
\value{
The autocorrelation measure of \code{df}.
If the input is not a data.frame, an error message is returned.
}
\description{
This function generates the autocorrelation measure of an data.frame object.
It is based on the durbin watson test. The test is based on detrended data.
The data.frame object should represent an multivariate ts. As input is
only required an object from the class data.frame.
Otherwise the function returns an error message.
}
\examples{
calculate_durbin_watson_test(df = dataframe_object,targetcol = 'var1')
}
