% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_mean}
\alias{calculate_mean}
\title{Generates the mean of the normalized values [0,1] of an ts object.}
\usage{
calculate_mean(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The normalized mean of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This is a function to generate the mean of the normalized values [0,1]
of an time series object. As input is only required an object from
the class time series. Otherwise the function returns an error message.
}
\examples{
calculate_mean(ts = tsobject)
}
