% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_partial_autocorrelation}
\alias{calculate_partial_autocorrelation}
\title{Generates the mean of the partial autocorrelationfunction (PACF)
of an ts object.}
\usage{
calculate_partial_autocorrelation(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The mean of the PACF of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This is a function to generate the mean of the PACF of an time series object.
As input is only required an object from the class time series.
Otherwise the function returns an error message.
}
\examples{
calculate_partial_autocorrelation(ts = tsobject)
}
