% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_periodicity}
\alias{calculate_periodicity}
\title{Generates the periodicity/frequency factor of an ts object.}
\usage{
calculate_periodicity(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The periodicity/frequency factor of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This function generates the periodicity/frequency factor of an
time series object. As input is only required an object from the
class time series. Otherwise the function returns an error message.
}
\examples{
calculate_periodicity(ts = tsobject)
}
