% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_quartile_distribution}
\alias{calculate_quartile_distribution}
\title{Generates the percentage of the values in the 4 quartiles of an ts object.}
\usage{
calculate_quartile_distribution(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The percentage of the values in the 4 quartiles of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This is a function to generate the percentage of the values in the 4
quartiles of an time series object. As input is only required an object from
the class time series. Otherwise the function returns an error message.
}
\examples{
calculate_quartile_distribution(ts = tsobject)
}
