% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_sd}
\alias{calculate_sd}
\title{Generates the standard deviation of an ts object.}
\usage{
calculate_sd(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The standard deviation of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This function generates the standard deviation of an time series object.
As input is only required an object from the class time series.
Otherwise the function returns an error message.
}
\examples{
calculate_sd(ts = tsobject)
}
