% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_selfsimilarity}
\alias{calculate_selfsimilarity}
\title{Generates the Hurst exponent of an ts object.}
\usage{
calculate_selfsimilarity(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The Hurst exponent of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This is a function to measure the self similarity by the Hurst exponent
of an time series object. As input is only required an object from the
class time series. Otherwise the function returns an error message.
}
\examples{
calculate_selfsimilarity(ts = tsobject)
}
