% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_skewness}
\alias{calculate_skewness}
\title{Generates the skewness of an ts object.}
\usage{
calculate_skewness(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The skewness of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This is a function to generate the skewness of an time series object.
As input is only required an object from the class time series.
Otherwise the function returns an error message.
}
\examples{
calculate_skewness(ts = tsobject)
}
