% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_stepchange_percentage}
\alias{calculate_stepchange_percentage}
\title{Generates the percentage of step changes of an ts object.}
\usage{
calculate_stepchange_percentage(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The percentage of step changes of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This is a function to generate the percentage of step changes of
an time series object. As input is only required an object from the class
time series. Otherwise the function returns an error message.
}
\examples{
calculate_stepchange_percentage(ts = tsobject)
}
