% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_turningpoint_percentage}
\alias{calculate_turningpoint_percentage}
\title{Generates the percentage of turning points of an ts object.}
\usage{
calculate_turningpoint_percentage(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The percentage of turning points of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This is a function to generate the percentage of turning points of
an time series object. Points are significant turning points if there
probability is smaller the significance level 5%.
As input is only required an object from the class time series.
Otherwise the function returns an error message.
}
\examples{
calculate_turningpoint_percentage(ts = tsobject)
}
