% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_taxonomy.R
\name{classify_ts}
\alias{classify_ts}
\title{Classifies a time series based on the defined ts taxonomy features.}
\usage{
classify_ts(ts, na_option = "random")
}
\arguments{
\item{ts}{Either a vector, time series or data.frame object representing
a time series is allowed.}

\item{na_option}{A string value containing either 'mean', 'random',
or'median'; Standard values is 'random'.}
}
\value{
The final list, containing all scaled feature factor values.
If the input \code{ts} is not a vector, ts or data.frame object
an error message is returned.
}
\description{
This is a function to calculate the different (statistical) features
from the defined time series taxonomy. The taxonomy can be applied to
classify univariate and multivariate time series based on it.
As input for \code{ts} is only required an object from the classes
time series, data.frame or vector. Each of them should represent an
time series. Otherwise the function returns an error message. Also, for
\code{na_option} is only required the string 'mean', 'median' or 'random'
allowed. This means, that all na values are either replaced by the mean,
median or a random number between the min and max of the ts.
The standard value of \code{na_option} is 'random'. The following 24
features: "number_of_observations", "number_of_attributes",
"coefficient of determination", "durbin watson test", "mean", "periodicity",
"chaos", "entropy", "selfsimilarity", "dynamic time warping (DTW) distance ",
"percentage of turning points", "variance", "percentage of outliers",
"percentage of step changes", "quartile distribition", "standard deviation",
"percentage of peaks", "trend", "seasonality", "autocorrelation",
"partial autocorrelation", "skewness", "kurtosis", "non linearity"
are calculated, scaled and then assigned to the taxonomy values.
}
\examples{
ts_vector = c(1,4,6,1,24,5,1)
classify_ts(ts = ts_vector_object)
classify_ts(ts = ts_vector_object, na_option = "random")
classify_ts(ts = dataframe_object, na_option = "mean")
classify_ts(ts = ts_object, na_option = "median")
}
