% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{decompose_ts}
\alias{decompose_ts}
\title{Decompose a time series object into trend, seasonal and remainder.}
\usage{
decompose_ts(ts, transform = TRUE)
}
\arguments{
\item{ts}{A time series object.}

\item{transform}{A boolean to enable BoxCox transformation.}
}
\value{
A list of the decomposed \code{ts}. If the input is not a ts,
an error message is returned. The list consist of the ts, the trend of
the ts, the seasonal of the ts and the remainder of the ts.
}
\description{
This is a function to decompose a time series object.
The time series object is decomposed into trend, seasonal and remainder.
Remainder is the ts minus the trend and seasonal.
As input is only required an object from the class time series.
Otherwise the function returns an error message.
}
\examples{
decompose_ts(ts = ts_object, transform = TRUE)
}
