% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_ts_frequency}
\alias{get_ts_frequency}
\title{Get the period of an vector object.}
\usage{
get_ts_frequency(x)
}
\arguments{
\item{x}{A vector object representing a time series.}
}
\value{
The period of \code{x}.
If the input is not a vector, an error message is returned.
}
\description{
This is a function to find the period of an vector object.
The vector object is representing a time series.
As input is only required an object from the class time series.
Otherwise the function returns an error message.
}
\examples{
ts_vector = c(1,4,6,1,24,5,1)
get_ts_frequency(x = ts_vector)
}
