% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matrix_block_list}
\alias{matrix_block_list}
\title{A list of 13 elements, containing time series matrices.}
\format{A list of 13 elements, each element is a matrix:
\describe{
  \item{stock}{A matrix, each row represents multivariate stock
  time series data from kaggle}
  \item{weather}{A matrix, each row represents multivariate weather
  time series data from kaggle}
  \item{crime}{A matrix, each row represents multivariate crime
  time series data from kaggle}
  \item{crypto}{A matrix, each row represents crypto currency
  time series data from kaggle}
  \item{multi-other}{A matrix, each row represents multivariate
  time series data of different kind}
  \item{finance}{A matrix, each row represents univariate finance
  time series data from the M4 competition}
  \item{micro}{A matrix, each row represents univariate micro
  time series data from the M4 competition}
  \item{demographic}{A matrix, each row represents univariate demographic
  time series data from the M4 competition}
  \item{macro}{A matrix, each row represents univariate macro
  time series data from the M4 competition}
  \item{industry}{A matrix, each row represents univariate industry
  time series data from the M4 competition}
  \item{other}{A matrix, each row represents univariate time series data
  of different kind from the M4 competition}
  \item{NN3}{A matrix, each row represents univariate time series data
  from the NN3 competition}
  \item{NNGC}{A matrix, each row represents univariate time series data from
  the NNGC competition}
}}
\usage{
matrix_block_list
}
\description{
A list of 13 elements, containing time series matrices for the dynamic time
warping distance calculation. The 1000 time series data from the
multi_ts_list.RDS and uni_ts_list.RDS files are divided into 13 blocks by
their data type.
}
\keyword{datasets}
