% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_ts_list}
\alias{multi_ts_list}
\title{A list of 500 different multivariate time series data.}
\format{A list with 500 elements.
Each element is a list of the following three items:
\describe{
  \item{name}{The valid name of the ts from its source}
  \item{data}{A data.frame which always contains a 'date' column and
  then one further column for the ts data}
  \item{desc}{A brief information about the kind of the data}
}}
\source{
Kaggle TS data \url{https://www.kaggle.com/datasets?tagids=6618}
University of Florida \url{http://users.stat.ufl.edu/~winner/datasets.html}
UCI Machine Learning Repository \url{https://archive.ics.uci.edu/ml/datasets.html?format=&task=reg&att=&area=&numAtt=&numIns=&type=ts&sort=attUp&view=table}
}
\usage{
multi_ts_list
}
\description{
A list containing 500 different multivariate time series data.
Each element of the \code{uni_list} represents a time series by a list
of three different attributes: name, data and desc. The name attribute
contains the name of the time series, the data attribute is a 2 to n
dimensional data frame consisting of the date and several actual time
series values/features. The desc attribute describes the kind of ts data.
458 elements of the list are multivariate time series data from the
kaggle data science platform. It consist mainly of stock, weather,
currency and crime data. The left 42 elemnts are from the University of
Florida and the UCI Machine Learning Repository.
}
\keyword{datasets}
