% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{predict_missing_observations}
\alias{predict_missing_observations}
\title{Replace all missing observations of a time series or vector object}
\usage{
predict_missing_observations(ts, na_option = "mean")
}
\arguments{
\item{ts}{A time series or vector object.}

\item{na_option}{A string value containing either 'mean'
or'kalman'; Standard values is 'mean'.}
}
\value{
The na imputated \code{ts}. If the input is not a ts or
vector, an error message is returned.
}
\description{
This is a function to Replace all missing observations of a time series or
vector object. For \code{na_option} is only required the string 'mean' or
'kalman' allowed. This means, that all na values are either replaced by the
mean, or kalman imputation of the ts. The standard value of \code{na_option}
is 'mean'. Otherwise the function returns an error message.
}
\examples{
predict_missing_observations(ts = datasets::BJsales)
}
