% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uni_ts_list}
\alias{uni_ts_list}
\title{A list of 500 different univariate time series data.}
\format{A list with 500 elements.
Each element is a list of the following three items:
\describe{
  \item{name}{The valid name of the ts from its source}
  \item{data}{A data.frame which always contains a 'date' column
  and then one further column for the ts data}
  \item{desc}{A brief information about the kind of the data}
}}
\source{
M4 competition \url{https://www.mcompetitions.unic.ac.cy/the-dataset/}
NNGC competition \url{http://www.neural-forecasting-competition.com/downloads/NNGC1/datasets/download.htm}
NN3 competition \url{http://www.neural-forecasting-competition.com/NN3/datasets.htm}
}
\usage{
uni_ts_list
}
\description{
A list containing 500 different univariate time series data.
Each element of the \code{uni_list} represents a time series by a list
of three different attributes: name, data and desc. The name attribute
contains the name of the time series, the data attribute is a 2 dimensional
data frame consisting of the date and the actual time series values. The
desc attribute describes the kind of time series data.
The first 323 elements of the list are time series data from the
M4 competition for forecasting model evaluation. Also, 66 elements of the
list are from the NNGC competition. The left 111 elements are time series
from the NN3 competition.
}
\keyword{datasets}
