% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{Generates the approximate entropy of an ts object.}
\usage{
calculate_entropy(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The approximate entropy of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This function generates the approximate entropy of an time series object.
As input is only required an object from the class time series.
Otherwise the function returns an error message.
}
\examples{
calculate_entropy(ts = tsobject)
}
