% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_non_linearity}
\alias{calculate_non_linearity}
\title{Generates the non linearity factor of an ts object.}
\usage{
calculate_non_linearity(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The non linearity factor of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This function generates the non linearity factor of an time series object.
The Teraesvirta's neural network test for neglected nonlinearity is applied.
As input is only required an object from the class time series.
Otherwise the function returns an error message.
}
\examples{
calculate_non_linearity(ts = tsobject)
}
