% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_outlier_percentage}
\alias{calculate_outlier_percentage}
\title{Generates the percentage of outliers of an ts object.}
\usage{
calculate_outlier_percentage(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The percentage of outliers of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This is a function to generate the percentage of outliers of an
time series object. Points are significant outliers if there probability is
smaller the significance level 5%. A z_test, chisq_test and t_test is
applied to identify number of outliers. Then the mean of the three test
results is used as final number of outliers. As input is only required an
object from the class time series. Otherwise the function returns
an error message.
}
\examples{
calculate_outlier_percentage(ts = tsobject)
}
