% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{calculate_peak_percentage}
\alias{calculate_peak_percentage}
\title{Generates the percentage of peaks of an ts object.}
\usage{
calculate_peak_percentage(ts)
}
\arguments{
\item{ts}{A time series object.}
}
\value{
The percentage of peaks of \code{ts}.
If the input is not a ts, an error message is returned.
}
\description{
This function generates the percentage of peaks of an time series object.
As input is only required an object from the class time series.
Otherwise the function returns an error message.
}
\examples{
calculate_peak_percentage(ts = tsobject)
}
